+++ lpp/sources/common.py
@@ -54,9 +54,9 @@ class TagSourceBase(ABC):
 
     def _send_api_request(
         self, endpoint: str, query_params: dict[str:str],
-        user_agent: str = "lazy-pony-prompter (by user Siberpone)/v1.1.1"
+        user_agent: str = "lazy-pony-prompter (by user Siberpone)/v1.1.x"
     ) -> dict[str:object]:
-        TIMEOUTS = (6.1, 9.1)
+        TIMEOUTS = (9.1, 15.1)
         req = requests.get(
             endpoint,
             query_params,