+++ lpp/sources/e621.py
@@ -122,3 +122,11 @@ class E621(TagSourceBase):
             .select("character", "species", "general", "artist", "meta")\
             .filter(self.filter)\
             .as_tag_groups()
+
+    @formatter(Models.NOOBAI.value)
+    def noobai_fomat(self, raw_image_tags: dict[str:list[str]]) -> TagGroups:
+        return Tags(self._convert_raw_tags(raw_image_tags))\
+            .select("character", "species", "general", "artist", "meta")\
+            .filter(self.filter)\
+            .replace_underscores()\
+            .as_tag_groups()