+++ scripts/lpp_extension.py
@@ -649,6 +649,7 @@ class Scripts(scripts.Script):
                 "461c3bbd5c": Models.SEAART.value,  # SeaArt Furry v1.0
                 "821aa5537f": Models.PDV56.value,   # AutismMix_pony
                 "ac006fdd7e": Models.PDV56.value,   # AutismMix_confetti
+                "ff827fc345": Models.NOOBAI.value   # NoobAI XL v1.0
             }
             if p.sd_model_hash not in model_hashes:
                 prompts_format = Models.PDV56.value